Machine Learning on USArrests.csv using Python
This data set contains statistics, in arrests per 100,000 residents for assault, murder, and rape in each of the 50 US states in 1973. 
Also given is the percent of the population living in urban areas.
A data frame with 50 observations on 4 variables namely; murder, assault, urban pop, rape.
Retrieved from World Almanac and Book of facts 1975. (Crime rates)
The goal of the project like all data analytics projects is to provide return on investment. This return would be a predictive model for law enforcement dispatches, new safety policies in states, and informative models for investors in the real estate development sector.